+++ src/index.mjs
@@ -1,4 +1,5 @@
 import {CAUSE, STRING, NUMERIC_STRING} from "./libs/constants";
+import AdjusterError from "./libs/AdjusterError";
 import adjust from "./libs/adjust";
 import boolean from "./adjusters/boolean";
 import number from "./adjusters/number";
@@ -12,6 +13,8 @@ export default {
 	CAUSE,
 	STRING,
 	NUMERIC_STRING,
+
+	AdjusterError,
 
 	adjust,