+++ test/index.test.mjs
@@ -91,6 +91,7 @@ function testError()
 				return;
 			}
 
+			expect(err instanceof adjuster.AdjusterError).toEqual(true);
 			switch(err.keyStack.shift())
 			{
 			case "id":